+++ CHANGELOG.rst
@@ -31,19 +31,51 @@ Change Log
 
 Unreleased
 ----------
+.. scriv-insert-here
 
-[2.1.0] - 2025-04-23
+[3.4.0] - 2026-05-07
 
 Added
 ~~~~~
 
 * Added django52 support.
 * Also releasing pending items.
+* Added new ``InstructorDashboardTabsRequested`` filter to allow for dynamic generation of instructor dashboard tabs. (by @holaontiveros)
 
 See the fragment files in the `changelog.d directory`_.
 .. _changelog.d directory: https://github.com/openedx/openedx-filters/tree/master/changelog.d
-.. scriv-insert-here
 
+[3.3.0] - 2025-04-17
+--------------------
+
+Changed
+~~~~~~~
+
+* Expand Python version compatibility
+
+[3.2.0] - 2025-04-13
+--------------------
+
+Changed
+~~~~~~~
+
+* Added GradeEventContextRequested filter
+
+[3.1.0] - 2025-04-06
+--------------------
+
+Changed
+~~~~~~~
+
+* Added AccountSettingsReadOnlyFieldsRequested filter
+
+[3.0.0] - 2025-03-19
+--------------------
+
+Changed
+~~~~~~~
+
+* Dropped Python 3.11 support
 
 [2.0.1] - 2025-02-18
 --------------------